feat: additional tests and refactorings
